Add warmup first iteration
The first iteration is slow, likely because of the overhead of thread creation,
which seems to impact MPI (both impi/2018.4 and mvapich2/2.3rc1) by increasing
the latency on some communications near the beginning of the calculation (to
several milliseconds). Extra threads are needed for the weak tasks, which stay
alive for autowait. Note: we want them to stay alive also so that there is more
opportunity to connect later tasks in the namespace.