Improve submit scripts.
The submit scripts does not use all the executable files in the current
directory based on a wildcard. Now the executable files need to be
passes as REST arguments at the end of the submission line.

Extra checks are added now to assert that the files are all executable
and that the REST list is not empty.

The output was also improved to print the submission command and a
logfile is now created in the output directory to check for errors in
the submission moment.

A new optional command line option was added to specify the output
directory (default: results)

The submit script also perform some checks.